Pause Point
Managed to implement deletion of TodoItems, but addition of TodoItems is missing POSTing of data to backend. This is because I don't know how to handle POSTing of e.g. module data, todo items data, when sending data to the backend. Do I send the ID of the module data and todoitems data? Or do I send the field (eg. "CS1101S") itself?